Fix timestamp overflow for videos longer than 13.5 hours
Change duration/position/resume fields from int (32-bit) to long (64-bit)
throughout the codebase to support videos longer than 24.8 days. This fixes
crashes when playing long-form content (20+ hours) where audio timestamps
exceed int32_max (2,147,483,647 milliseconds).

Root Cause:
- Audio timestamps at 44.1kHz sample rate overflow int32 for videos >13.5 hours
- Example: 20-hour video has duration_ts = 3,297,391,616 (exceeds int32_max)
- This causes negative values, array index errors, and crashes

Changes Made:

Core Data Model:
- VideoMetadata: duration field int -> long
- Video: mDurationMs, mResumeMs, mRemoteResume int -> long
- VideoProperties: duration, resumePosition, bookmarkPosition int -> long
- Player: mDuration and position fields int -> long

Database Layer:
- VideoCursorMapper: cursor.getInt() -> cursor.getLong() for duration/resume
- AnimesNShowsMapper: cursor.getInt() -> cursor.getLong() for duration/resume

Player/Controllers:
- Player: getDuration(), getCurrentPosition(), seekTo() use long
- PlayerService: duration/position variables int -> long
- PlayerController: duration/position variables int -> long
- FloatingPlayerService: duration/position variables int -> long

UI/Presenters:
- VideoPresenter: setResume() parameters int -> long
- VideoGridPresenter: duration/resume variables int -> long
- VideoListPresenter: duration/resume variables int -> long
- EpisodePresenter: duration/resume variables int -> long

Files Modified: 14 files
Lines Changed: ~120 changes across codebase

Testing:
- Java long is 64-bit signed: supports durations up to 292 million years
- SQLite INTEGER type already supports 64-bit values (no migration needed)
- Android MediaPlayer API supports long for duration/position
- Changes maintain backward compatibility (long accepts int values)

Fixes: #1615

+++ src/main/java/com/archos/mediacenter/video/browser/VideoProperties.java
@@ -31,11 +31,11 @@ public class VideoProperties {
         return bookmark;
     }
 
-    public int getBookmarkPosition() {
+    public long getBookmarkPosition() {
         return bookmarkPosition;
     }
 
-    public int getResumePosition() {
+    public long getResumePosition() {
         return resumePosition;
     }
 
@@ -43,7 +43,7 @@ public class VideoProperties {
         return subtitles;
     }
 
-    public int getDuration() {
+    public long getDuration() {
         return duration;
     }
 
@@ -109,7 +109,7 @@ public class VideoProperties {
         this.bookmark = bookmark;
     }
 
-    public void setBookmarkPosition(int bookmarkPosition) {
+    public void setBookmarkPosition(long bookmarkPosition) {
         this.bookmarkPosition = bookmarkPosition;
     }
 
@@ -117,7 +117,7 @@ public class VideoProperties {
         this.resume = resume;
     }
 
-    public void setResumePosition(int resumePosition) {
+    public void setResumePosition(long resumePosition) {
         this.resumePosition = resumePosition;
     }
 
@@ -125,7 +125,7 @@ public class VideoProperties {
         this.subtitles = subtitles;
     }
 
-    public void setDuration(int duration) {
+    public void setDuration(long duration) {
         this.duration = duration;
     }
 
@@ -184,11 +184,11 @@ public class VideoProperties {
     public void setNameList(String nameList) {
         this.nameList = nameList;
     }
-    public int bookmarkPosition;
+    public long bookmarkPosition;
     public boolean resume;
-    public int resumePosition;
+    public long resumePosition;
     public boolean subtitles;
-    public int duration;
+    public long duration;
     public int id;
     public int scraperType;
     public float rating;
@@ -200,8 +200,8 @@ public class VideoProperties {
     public String detailLineOne, detailLineTwo, detailLineThree;
     public String name, nameGrid, nameList;
 
-    public VideoProperties(int id, int duration, boolean bookmark, int bookmarkPosition,
-                           boolean resume, int resumePosition, boolean subtitles,
+    public VideoProperties(int id, long duration, boolean bookmark, long bookmarkPosition,
+                           boolean resume, long resumePosition, boolean subtitles,
                            int scraperType, String name, String nameGrid, String nameList,
                            float rating, String detailLineOne, String detailLineTwo,
                            String detailLineThree, String coverPath, long date, int traktSeen,
