+++ src/main/java/com/archos/mediacenter/video/browser/presenter/VideoGridPresenter.java
@@ -76,10 +76,10 @@ public class VideoGridPresenter extends VideoPresenter{
 
 
 
-        int resumePosition = video.getRemoteResumeMs()>0?video.getRemoteResumeMs():video.getResumeMs();
+        long resumePosition = video.getRemoteResumeMs()>0?video.getRemoteResumeMs():video.getResumeMs();
         boolean resume = resumePosition>0||resumePosition == PlayerActivity.LAST_POSITION_END;
         if (resume&&holder.resume!=null) {
-            int duration = video.getDurationMs();
+            long duration = video.getDurationMs();
             duration = duration > 0 ? duration : resumePosition>0&&resumePosition<=100? 100 : 0;//resume can now be a percentage
             boolean displayProgressSlider = !mIsTablet&&(duration>0 ||resumePosition == PlayerActivity.LAST_POSITION_END); // Display the progress bar if we know the duration
             setResume(displayProgressSlider,duration > 0 ? duration : 100, resumePosition, holder.resume);
