+++ src/main/java/com/archos/mediacenter/video/browser/presenter/VideoListPresenter.java
@@ -88,10 +88,10 @@ public class VideoListPresenter extends VideoPresenter{
         }
         if(holder.name!=null)
             holder.name.setEllipsize(TextUtils.TruncateAt.END);
-        int resumePosition = video.getRemoteResumeMs()>0?video.getRemoteResumeMs():video.getResumeMs();
+        long resumePosition = video.getRemoteResumeMs()>0?video.getRemoteResumeMs():video.getResumeMs();
         boolean resume = resumePosition>0 || resumePosition == PlayerActivity.LAST_POSITION_END;
         if (resume&&holder.resume!=null) {
-            int duration = video.getDurationMs();
+            long duration = video.getDurationMs();
             duration = duration > 0 ? duration : resumePosition>0&&resumePosition<=100? 100 : 0;//resume can now be a percentage
             boolean displayProgressSlider = !mThinPhoneInPortrait&&(duration>0 ||resumePosition == PlayerActivity.LAST_POSITION_END); // Display the progress bar if we know the duration
             setResume(displayProgressSlider,duration > 0 ? duration : 100, resumePosition, holder.resume);
